improved naming (see rust naming convention)
