+++ src/messages/mod.rs
@@ -1,4 +1,4 @@
-pub mod fragmentrequest;
-pub mod fragmenttask;
-pub mod fragmentresult;
+pub mod fragment_request;
+pub mod fragment_task;
+pub mod fragment_result;
 pub mod complementary_types;
